Add files via upload
This is an R script to identify the co-expressed genes and hub genes from the RNASeq expression data using the WGCNA tool from R code. The expression data should be normalized first and the adjacency matrix is calculated for each gene with every gene. The output will provide a csv file with a gene-gene connection with a weightage value which indicates the strength of the co-expression. The output from the code can be used to construct the network using the Cytoscape tool by uploading the file to the software.